fix float numbers valueble

+++ js/task-3.js
@@ -1,7 +1,7 @@
 'use strict';
 
 function getElementWidth(content, padding, border) {
-    let totalWidth = Number.parseInt(content) + Number.parseInt(padding) * 2 + Number.parseInt(border) * 2;
+    let totalWidth = Number.parseFloat(content) + Number.parseFloat(padding) * 2 + Number.parseFloat(border) * 2;
 
     return totalWidth;
 }
